Performance tweak for Mysql 4.0
Mysql 4.0 does not support Sqlmode command, in addition to that
the old provider does not support the server version property.

To compensate for this Simple.Data.Mysql have so far used try / catch
to attempt to get the data, and this was not cached in any way.

This commit attempts to remedy this by maintaining a list og connection
strings that causes exception, and subsequently skip the try catch
for these connections strings.