Schemadata strategy selection in place
Different schemadata providers are now used depending on what server
version is used. This way the provider can dynamically select how the
schema data is retrieved, and also one can implement other details
depending on how the server expects it.

Quoting of object names is so far the only difference in implementation
of 4.x strategy and 5.x strategy. (Closes issue #22)